New endpoints to update and create single entities
Problem
-------
to make the editor a bit easier to deal with, we need to have endpoints that
will simply take the entity as a hash and insert or updated it into the gamestate.

Solution
--------

* New Routes to new Api.EntitiesController
* build update and create endpoints
* moved behavior atomizing/finding into the Behavior module
* added Doro.TestFixtures.read_fixture helper

Notes
-----

I feel like some of this is messy (especially the tests) and I'm not sure
if it's that I don't know how to use Elixir or if i'm missing division
lines about where to put functions/modules etc.

Please comment on style/organization if you can see improvements.