Get edit working from the game console
Problem
-------
We reworked the structure of Behaviors and we needed the front end to
understand that update

Solution
--------

In react, we take the incoming marshalled behaviors and put them (almost
directly) into the state with the small change of mapping behavior
arrays from `[{type: 'visible', fields...}]` to
`{ visible: { fields... } }`.  This makes it a bit easier to keep track
of changes in the local state.

When we send the data to the server (via the `addEntitySaga`) we convert
the behaviors back and submit to the server.

Changes
-------

* Lots of refactoring.
* Trying to move away from ramda
* add and edit actions both go through the same endpoint -
`EntitiesController::create`
* make the command `/edit whatever` open up the editor pane and properly
populate the form
* added a few fixes to the world data where we'd either forgotten or
merged in old style behaviors